Lots of themes added (2), design work, product addition now a thing (s3 not completed, wrong url)

+++ lib/tasks/s3_import.rake
@@ -1,3 +1,5 @@
+require "net/http"
+
 namespace :s3_import do
   desc "Looks at all products in Product table and creates
 	image attachment url for s3 based on photo directory"
@@ -15,18 +17,24 @@ namespace :s3_import do
                 url = "http://www.vigme.com/photos/"
                 path = product.image_folder.to_s + "/" + product.id.to_s + ".jpg"
                 url = url + path
-                product.image_s3 = url
-                # add datetimes and delete old images
-				product.ftp_transfer_processed = 1
-				product.ftp_transfer_datetime = Time.now
-				product.image_s3_url = product.image_s3.url
-				product.ftp_transfer_deleted_source = 1
-				product.ftp_transfer_deleted_source_datetime = Time.now
-				product.save
-				p "Product Id:   " + product.id.to_s
-				p "Amazon s3 URL:" + product.image_s3.url.to_s
-				path_to_file = "/var/www/vigme.com/photos/" + path
-				File.delete(path_to_file) if File.exist?(path_to_file)
+                p url 
+                urlCheck = URI.parse(url)
+                req = Net::HTTP.new(urlCheck.host, urlCheck.port)
+                res = req.request_head(urlCheck.path)
+                if res.code != "404"
+		            product.image_s3 = url
+		            # add datetimes and delete old images
+					product.ftp_transfer_processed = 1
+					product.ftp_transfer_datetime = Time.now
+					product.image_s3_url = product.image_s3.url
+					product.ftp_transfer_deleted_source = 1
+					product.ftp_transfer_deleted_source_datetime = Time.now
+					product.save
+					p "Product Id:   " + product.id.to_s
+					p "Amazon s3 URL:" + product.image_s3.url.to_s
+					path_to_file = "/var/www/vigme.com/photos/" + path
+					File.delete(path_to_file) if File.exist?(path_to_file)
+				end
         end
 	    pageNumber = pageNumber + 1
 	  else
